feat: 접속 페이지 디자인 Intra 로그인 할때만 표시, 첫 화면 메인페이지 id ===0일때 문구 변경
